requirements.txt
aiosqlite==0.17.0
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncmy==0.2.9
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
iso8601==1.1.0
kombu==5.3.5
prompt-toolkit==3.0.43
pydantic==2.6.4
pydantic_core==2.16.3
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.3
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tortoise-orm==0.20.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.7.0
